+++ src/pages/MyspacePage.tsx
@@ -1,42 +1,39 @@
 import { useState, useEffect } from 'react';
 import { useLocation } from 'react-router-dom';
 import Header from '@components/Header';
-import Modal from '@components/Modal';
 import Card from '@components/Card';
 import SearchSVG from '@assets/search.svg?react';
 import ArrowSVG from '@assets/arrow.svg?react';
-
+import useModal from '@/hooks/useModal';
+import LoginModal from '@/components/LoginModal';
+import SurveyModal from '@/components/SurveyModal';
 type form = { id: string; title: string; content: string; tagList: string[] };
 
 const MyspacePage = () => {
-  const [modalOpen, setModalOpen] = useState<boolean>(false);
-  const [modalType, setModalType] = useState<string>('survey');
   const [surveyDataList, setSurveyDataList] = useState<form[]>([]);
   const [surveyData, setSurveyData] = useState<form>({ id: '', title: '', content: '', tagList: [] });
-
   const location = useLocation();
   const nickName = location.state ? location.state.nickName : '';
+  const { modalType, modalOpen, openModal, closeModal, formData, handleFormData } = useModal();
+  useEffect(() => {
+    if (formData.id !== '') {
+      setSurveyDataList((prevList) => [...prevList, formData]); // 이전 목록에 새로운 데이터 추가
+    }
+  }, [formData]);
 
-  const receiveFormData = (data: form) => {
-    const list = surveyDataList;
-    list?.push(data);
-    console.log(list);
-
-    setSurveyDataList(list);
-    console.log(surveyDataList);
-    console.log('Received form data:' + data.id + ' ' + data.title);
-  };
-
-  const showModal = (type: string) => {
-    setModalType(type);
-    setModalOpen(true);
-  };
   return (
     <div className="h-screen w-full bg-[#FAFAFA]">
       <Header headText={'마이스페이스 💌'} nickName={`${nickName[0]}`} />
       <div className="relative top-[5.6rem] h-[8rem] border-b border-[#E5E5E5] bg-white px-[3.2rem]">
         <button
-          onClick={() => showModal(modalType)}
+          onClick={() =>
+            openModal('survey', {
+              id: '',
+              title: '',
+              content: '',
+              tagList: [],
+            })
+          }
           className="mb-[0.78rem] h-[3.6rem] w-[13.2rem] rounded-[0.4rem] bg-[#262626] text-[1.4rem] font-bold text-white opacity-20 hover:opacity-100"
         >
           새 설문 만들기
@@ -60,18 +57,17 @@ const MyspacePage = () => {
       </div>
       <div className="px-[3.2rem]">
         <div className="relative top-[9rem] grid grid-cols-1 gap-4 md:grid-cols-2 xl:grid-cols-3 2xl:grid-cols-4 ">
-          {surveyDataList?.map((data) => (
+          {surveyDataList.map((data) => (
             <Card key={data.id} id={data.id} title={data.title} content={data.content} tagList={data.tagList} />
           ))}
         </div>
       </div>
       {modalOpen && (
-        <Modal
-          modalType={modalType}
-          setModalOpen={setModalOpen}
-          formData={surveyData}
-          receiveFormData={receiveFormData}
-        />
+        <div className="modal-overlay">
+          {modalType === 'survey' && (
+            <SurveyModal setModalOpen={closeModal} formData={formData} handleFormData={handleFormData} />
+          )}
+        </div>
       )}
     </div>
   );
